requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector-python==8.0.28
oauthlib==3.2.0
protobuf==3.19.4
requests==2.27.1
requests-oauthlib==1.3.1
soupsieve==2.3.1
tweepy==4.8.0
urllib3==1.26.9
